model/requirements.txt
diambra-arena==2.2.7
stable-baselines3[extra]
torch
numpy

